mediacapture-streams/MediaStream-supported-by-feature-policy.html

<!DOCTYPE html>
<title>Test that cemara and microphone are advertised in the permission list</title>
<link rel="help" href="https://w3c.github.io/webappsec-feature-policy/#dom-featurepolicy-features">
<link rel="help" href="https://w3c.github.io/mediacapture-main/#feature-policy-integration">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('camera', document.permissionsPolicy.features());
}, 'document.permissionsPolicy.features should advertise camera.');

test(() => {
    assert_in_array('microphone', document.permissionsPolicy.features());
}, 'document.permissionsPolicy.features should advertise microphone.');
</script>
